feat: replace create with prepare
