+++ src/logic/create/map.test.ts
@@ -1,12 +1,19 @@
 import { EXTENSIONS } from '@superfaceai/ast';
 
-import { getProfileFile, MockLogger } from '../../common';
+import { getProfileFile, getProviderFile, MockLogger } from '../../common';
 import { createUserError } from '../../common/error';
+import { readFile } from '../../common/io';
 import { OutputStream } from '../../common/output-stream';
 import { ProfileId } from '../../common/profile';
 import { mockProfileDocumentNode } from '../../test/profile-document-node';
-import { prepareMockMap } from './mock-map';
+import { mockProviderJson } from '../../test/provider-json';
+import { createMap } from './map';
 import { loadProfileAst } from './utils';
+
+jest.mock('../../common/io', () => ({
+  ...jest.requireActual('../../common/io'),
+  readFile: jest.fn(),
+}));
 
 jest.mock('./utils', () => ({
   loadProfileAst: jest.fn(),
@@ -14,15 +21,17 @@ jest.mock('./utils', () => ({
 
 jest.mock('../../common/super-json-utils', () => ({
   getProfileFile: jest.fn(),
+  getProviderFile: jest.fn(),
 }));
 
-describe('Prepare mock map logic', () => {
+describe('Prepare map logic', () => {
   let logger: MockLogger;
   const userError = createUserError(false);
   const profileId = ProfileId.fromScopeName('test', 'profile');
   const profilePath = 'profilePath';
 
   const provider = 'provider';
+  const providerPath = 'providerPath';
 
   const mockProfile = mockProfileDocumentNode({
     name: profileId.name,
@@ -43,6 +52,7 @@ describe('Prepare mock map logic', () => {
   };
 
   const superJsonPath = 'path/to/super.json';
+  const mockProvider = mockProviderJson({ name: provider });
 
   const writeOnceSpy = jest.spyOn(OutputStream, 'writeOnce');
 
@@ -50,7 +60,10 @@ describe('Prepare mock map logic', () => {
     writeOnceSpy.mockResolvedValue(undefined);
 
     jest.mocked(getProfileFile).mockResolvedValue(profilePath);
+    jest.mocked(getProviderFile).mockResolvedValue(providerPath);
+
     jest.mocked(loadProfileAst).mockResolvedValue(mockProfile);
+    jest.mocked(readFile).mockResolvedValue(JSON.stringify(mockProvider));
 
     logger = new MockLogger();
   });
@@ -59,31 +72,8 @@ describe('Prepare mock map logic', () => {
     jest.resetAllMocks();
   });
 
-  it('Writes prepared mock map to file', async () => {
-    await prepareMockMap(
-      {
-        id: {
-          profile: profileId,
-        },
-        superJson: mockSuperJson,
-        superJsonPath,
-      },
-      {
-        logger,
-        userError,
-      }
-    );
-
-    expect(writeOnceSpy).toBeCalledWith(
-      `${profileId.toString()}.mock${EXTENSIONS.map.source}`,
-      expect.any(String),
-      { dirs: true, force: undefined }
-    );
-    expect(writeOnceSpy).toBeCalledWith(superJsonPath, expect.any(String));
-  });
-
-  it('Writes prepared mock map to file with provider name', async () => {
-    await prepareMockMap(
+  it('Writes prepared map to file', async () => {
+    await createMap(
       {
         id: {
           profile: profileId,
@@ -101,16 +91,42 @@ describe('Prepare mock map logic', () => {
     expect(writeOnceSpy).toBeCalledWith(
       `${profileId.toString()}.${provider}${EXTENSIONS.map.source}`,
       expect.any(String),
+      { dirs: true, force: undefined }
+    );
+    expect(writeOnceSpy).toBeCalledWith(superJsonPath, expect.any(String));
+  });
+
+  it('Writes prepared map to file with variant', async () => {
+    await createMap(
+      {
+        id: {
+          profile: profileId,
+          provider,
+          variant: 'test',
+        },
+        superJson: mockSuperJson,
+        superJsonPath,
+      },
+      {
+        logger,
+        userError,
+      }
+    );
+
+    expect(writeOnceSpy).toBeCalledWith(
+      `${profileId.toString()}.${provider}.test${EXTENSIONS.map.source}`,
+      expect.stringContaining('variant = "test"'),
       { dirs: true }
     );
     expect(writeOnceSpy).toBeCalledWith(superJsonPath, expect.any(String));
   });
 
-  it('Writes prepared mock map to file with force flag', async () => {
-    await prepareMockMap(
+  it('Writes prepared map to file with force flag', async () => {
+    await createMap(
       {
         id: {
           profile: profileId,
+          provider,
         },
         superJson: mockSuperJson,
         superJsonPath,
@@ -125,18 +141,19 @@ describe('Prepare mock map logic', () => {
     );
 
     expect(writeOnceSpy).toBeCalledWith(
-      `${profileId.toString()}.mock${EXTENSIONS.map.source}`,
+      `${profileId.toString()}.${provider}${EXTENSIONS.map.source}`,
       expect.any(String),
       { dirs: true, force: true }
     );
     expect(writeOnceSpy).toBeCalledWith(superJsonPath, expect.any(String));
   });
 
-  it('Writes prepared mock map to file with station flag', async () => {
-    await prepareMockMap(
+  it('Writes prepared map to file with station flag', async () => {
+    await createMap(
       {
         id: {
           profile: profileId,
+          provider,
         },
         superJson: mockSuperJson,
         superJsonPath,
@@ -152,7 +169,7 @@ describe('Prepare mock map logic', () => {
     );
 
     expect(writeOnceSpy).toBeCalledWith(
-      `grid/${profileId.id}/maps/mock${EXTENSIONS.map.source}`,
+      `grid/${profileId.id}/maps/${provider}${EXTENSIONS.map.source}`,
       expect.any(String),
       { dirs: true, force: false }
     );
