+++ src/logic/create/map.ts
@@ -1,17 +1,18 @@
 import type { SuperJsonDocument } from '@superfaceai/ast';
-import { EXTENSIONS } from '@superfaceai/ast';
+import { assertProviderJson, EXTENSIONS } from '@superfaceai/ast';
 import { mergeProfileProvider, NodeFileSystem } from '@superfaceai/one-sdk';
 
 import type { ILogger } from '../../common';
-import { getProfileFile } from '../../common';
+import { getProfileFile, getProviderFile } from '../../common';
 import type { UserError } from '../../common/error';
+import { readFile } from '../../common/io';
 import { OutputStream } from '../../common/output-stream';
 import { resolveSuperfaceRelativePath } from '../../common/path';
 import type { ProfileId } from '../../common/profile';
-import { prepareMockMapTemplate } from '../../templates/prepared-map';
+import { prepareMapTemplate } from '../../templates/prepared-map';
 import { loadProfileAst } from './utils';
 
-export async function prepareMockMap(
+export async function createMap(
   {
     id,
     superJson,
@@ -20,7 +21,8 @@ export async function prepareMockMap(
   }: {
     id: {
       profile: ProfileId;
-      provider?: string;
+      provider: string;
+      variant?: string;
     };
     superJson: SuperJsonDocument;
     superJsonPath: string;
@@ -40,22 +42,29 @@ export async function prepareMockMap(
   );
   const profileAst = await loadProfileAst(profileFile, { userError });
 
-  const mockMapTemplate = prepareMockMapTemplate(profileAst);
+  // Load provider
+  const providerFile = await getProviderFile(
+    id.provider,
+    { superJson, superJsonPath },
+    { userError }
+  );
+  const provider = assertProviderJson(
+    JSON.parse(await readFile(providerFile, { encoding: 'utf-8' }))
+  );
+
+  const mapTemplate = prepareMapTemplate(profileAst, provider, id.variant);
 
   // Write result
   let filePath: string;
+  const variantName = id.variant !== undefined ? `.${id.variant}` : '';
 
   if (options?.station === true) {
-    filePath = `grid/${id.profile.id}/maps/${id.provider ?? 'mock'}${
-      EXTENSIONS.map.source
-    }`;
+    filePath = `grid/${id.profile.id}/maps/${id.provider}${EXTENSIONS.map.source}`;
   } else {
-    filePath = `${id.profile.id}.${id.provider ?? 'mock'}${
-      EXTENSIONS.map.source
-    }`;
+    filePath = `${id.profile.id}.${id.provider}${variantName}${EXTENSIONS.map.source}`;
   }
 
-  const created = await OutputStream.writeIfAbsent(filePath, mockMapTemplate, {
+  const created = await OutputStream.writeIfAbsent(filePath, mapTemplate, {
     force: options?.force,
     dirs: true,
   });
@@ -64,7 +73,7 @@ export async function prepareMockMap(
     logger.success(
       'prepareMap',
       id.profile.id,
-      id.provider ?? 'mock',
+      id.provider,
       filePath,
       options?.station
     );
@@ -72,7 +81,7 @@ export async function prepareMockMap(
     mergeProfileProvider(
       superJson,
       id.profile.id,
-      id.provider ?? 'mock',
+      id.provider,
       {
         file: resolveSuperfaceRelativePath(superJsonPath, filePath),
       },
