+++ src/logic/create/mock-map.test.ts
@@ -1,19 +1,12 @@
 import { EXTENSIONS } from '@superfaceai/ast';
 
-import { getProfileFile, getProviderFile, MockLogger } from '../../common';
+import { getProfileFile, MockLogger } from '../../common';
 import { createUserError } from '../../common/error';
-import { readFile } from '../../common/io';
 import { OutputStream } from '../../common/output-stream';
 import { ProfileId } from '../../common/profile';
 import { mockProfileDocumentNode } from '../../test/profile-document-node';
-import { mockProviderJson } from '../../test/provider-json';
-import { prepareMap } from './map';
+import { createMockMap } from './mock-map';
 import { loadProfileAst } from './utils';
-
-jest.mock('../../common/io', () => ({
-  ...jest.requireActual('../../common/io'),
-  readFile: jest.fn(),
-}));
 
 jest.mock('./utils', () => ({
   loadProfileAst: jest.fn(),
@@ -21,17 +14,15 @@ jest.mock('./utils', () => ({
 
 jest.mock('../../common/super-json-utils', () => ({
   getProfileFile: jest.fn(),
-  getProviderFile: jest.fn(),
 }));
 
-describe('Prepare map logic', () => {
+describe('Prepare mock map logic', () => {
   let logger: MockLogger;
   const userError = createUserError(false);
   const profileId = ProfileId.fromScopeName('test', 'profile');
   const profilePath = 'profilePath';
 
   const provider = 'provider';
-  const providerPath = 'providerPath';
 
   const mockProfile = mockProfileDocumentNode({
     name: profileId.name,
@@ -52,7 +43,6 @@ describe('Prepare map logic', () => {
   };
 
   const superJsonPath = 'path/to/super.json';
-  const mockProvider = mockProviderJson({ name: provider });
 
   const writeOnceSpy = jest.spyOn(OutputStream, 'writeOnce');
 
@@ -60,10 +50,7 @@ describe('Prepare map logic', () => {
     writeOnceSpy.mockResolvedValue(undefined);
 
     jest.mocked(getProfileFile).mockResolvedValue(profilePath);
-    jest.mocked(getProviderFile).mockResolvedValue(providerPath);
-
     jest.mocked(loadProfileAst).mockResolvedValue(mockProfile);
-    jest.mocked(readFile).mockResolvedValue(JSON.stringify(mockProvider));
 
     logger = new MockLogger();
   });
@@ -72,8 +59,31 @@ describe('Prepare map logic', () => {
     jest.resetAllMocks();
   });
 
-  it('Writes prepared map to file', async () => {
-    await prepareMap(
+  it('Writes prepared mock map to file', async () => {
+    await createMockMap(
+      {
+        id: {
+          profile: profileId,
+        },
+        superJson: mockSuperJson,
+        superJsonPath,
+      },
+      {
+        logger,
+        userError,
+      }
+    );
+
+    expect(writeOnceSpy).toBeCalledWith(
+      `${profileId.toString()}.mock${EXTENSIONS.map.source}`,
+      expect.any(String),
+      { dirs: true, force: undefined }
+    );
+    expect(writeOnceSpy).toBeCalledWith(superJsonPath, expect.any(String));
+  });
+
+  it('Writes prepared mock map to file with provider name', async () => {
+    await createMockMap(
       {
         id: {
           profile: profileId,
@@ -91,42 +101,16 @@ describe('Prepare map logic', () => {
     expect(writeOnceSpy).toBeCalledWith(
       `${profileId.toString()}.${provider}${EXTENSIONS.map.source}`,
       expect.any(String),
-      { dirs: true, force: undefined }
-    );
-    expect(writeOnceSpy).toBeCalledWith(superJsonPath, expect.any(String));
-  });
-
-  it('Writes prepared map to file with variant', async () => {
-    await prepareMap(
-      {
-        id: {
-          profile: profileId,
-          provider,
-          variant: 'test',
-        },
-        superJson: mockSuperJson,
-        superJsonPath,
-      },
-      {
-        logger,
-        userError,
-      }
-    );
-
-    expect(writeOnceSpy).toBeCalledWith(
-      `${profileId.toString()}.${provider}.test${EXTENSIONS.map.source}`,
-      expect.stringContaining('variant = "test"'),
       { dirs: true }
     );
     expect(writeOnceSpy).toBeCalledWith(superJsonPath, expect.any(String));
   });
 
-  it('Writes prepared map to file with force flag', async () => {
-    await prepareMap(
+  it('Writes prepared mock map to file with force flag', async () => {
+    await createMockMap(
       {
         id: {
           profile: profileId,
-          provider,
         },
         superJson: mockSuperJson,
         superJsonPath,
@@ -141,19 +125,18 @@ describe('Prepare map logic', () => {
     );
 
     expect(writeOnceSpy).toBeCalledWith(
-      `${profileId.toString()}.${provider}${EXTENSIONS.map.source}`,
+      `${profileId.toString()}.mock${EXTENSIONS.map.source}`,
       expect.any(String),
       { dirs: true, force: true }
     );
     expect(writeOnceSpy).toBeCalledWith(superJsonPath, expect.any(String));
   });
 
-  it('Writes prepared map to file with station flag', async () => {
-    await prepareMap(
+  it('Writes prepared mock map to file with station flag', async () => {
+    await createMockMap(
       {
         id: {
           profile: profileId,
-          provider,
         },
         superJson: mockSuperJson,
         superJsonPath,
@@ -169,7 +152,7 @@ describe('Prepare map logic', () => {
     );
 
     expect(writeOnceSpy).toBeCalledWith(
-      `grid/${profileId.id}/maps/${provider}${EXTENSIONS.map.source}`,
+      `grid/${profileId.id}/maps/mock${EXTENSIONS.map.source}`,
       expect.any(String),
       { dirs: true, force: false }
     );
